Move localization classes back to package
