+++ common/app/com/commercetools/sunrise/framework/localization/NoLocaleFoundException.java
@@ -1,4 +1,4 @@
-package com.commercetools.sunrise.ctp.project;
+package com.commercetools.sunrise.framework.localization;
 
 public class NoLocaleFoundException extends RuntimeException {
