+++ common/app/com/commercetools/sunrise/framework/localization/UserLanguage.java
@@ -1,4 +1,4 @@
-package com.commercetools.sunrise.ctp.project;
+package com.commercetools.sunrise.framework.localization;
 
 import com.google.inject.ImplementedBy;
