+++ common/app/com/commercetools/sunrise/framework/template/engine/EmailHtmlContentRenderer.java
@@ -1,7 +1,7 @@
 package com.commercetools.sunrise.framework.template.engine;
 
 import com.commercetools.sunrise.cms.CmsService;
-import com.commercetools.sunrise.ctp.project.UserLanguage;
+import com.commercetools.sunrise.framework.localization.UserLanguage;
 import com.commercetools.sunrise.framework.viewmodels.PageData;
 import com.commercetools.sunrise.framework.viewmodels.content.PageContent;
 import play.twirl.api.Content;
