+++ common/conf/reference.conf
@@ -412,7 +412,7 @@ play = {
     enabled += "com.commercetools.sunrise.framework.injection.SunriseModule"
     enabled += "com.commercetools.sunrise.ctp.client.SphereClientModule"
     enabled += "com.commercetools.sunrise.ctp.project.ProjectModule"
-    enabled += "com.commercetools.sunrise.ctp.project.LocalizationModule"
+    enabled += "com.commercetools.sunrise.framework.localization.LocalizationModule"
     enabled += "com.commercetools.sunrise.framework.template.TemplateModule"
   }
 }
