+++ common/test/com/commercetools/sunrise/ctp/project/ProjectContextTest.java
@@ -1,5 +1,8 @@
 package com.commercetools.sunrise.ctp.project;
 
+import com.commercetools.sunrise.framework.localization.NoCountryFoundException;
+import com.commercetools.sunrise.framework.localization.NoCurrencyFoundException;
+import com.commercetools.sunrise.framework.localization.NoLocaleFoundException;
 import com.neovisionaries.i18n.CountryCode;
 import org.junit.Test;
